==> src/3. GDP_processing.ipynb <==
# !git clone -b master https://github.com/StefanoNasini/Country-by-country-2022-2023-international-relationships
# !cd Country-by-country-2022-2023-international-relationships

import os
root = os.path.dirname(os.getcwd())

def parse_file(file_path):
    data_list = []
    with open(file_path, 'r') as file:
        for line in file:
            line_data = line.strip().split()
            country_name = ' '.join(line_data[1:-1])
            population = int(line_data[-1].replace(',', ''))
            data_list.append((country_name, population))
    return data_list

file_path = os.path.join(root, "raw_data", "GDP_data", "GDP_data.txt")
result_list = parse_file(file_path)
print(result_list)

import os

def find_root_directory():
    current_dir = os.getcwd()
    root_dir = os.path.dirname(current_dir)
    return root_dir

try:
    print("Root Directory:", root)
except NameError:
    root = find_root_directory()
    print("Root Directory:", root)

os.chdir(os.path.join(root, "src", "utils"))
from aka import tri2aka, guess_country

# change result list's format into trigraphs
GDP = {}
remains = []
for (name, gdp) in result_list:
    try: 
        tri = guess_country(name) # if a unique country trigraph can be identified
        GDP[tri] = gdp
    except :  
        remains.append((name, gdp))

remains

guess_country("sao tome", show_guess_country=True)

l = "CHN IND GBR KOR NLD TUR IRL NGA HKG ROU NZL OMN DOM CIV SDN PNG GEO MAC GIN MLI LAO NER GNQ not_found SOM CUW GNB WSM ASM DMA STP SSD COD" 

for i, x in enumerate(remains):
    print(x, l.split(' ')[i])

for (tri, gdp) in zip(l.split(" "), dict(result_list).values()):
    GDP[tri] = gdp

try:
    GDP.pop("not_found") # the channel island is not in any other parts of data. Need to drop it.
except:
    pass

with open(os.path.join(root, "raw_data", "GDP_data", "GDP_data.json"), "w") as f:
    f.write(str(GDP))

guess_country("democratic republic of congo")
