==> src/4a. WTO_trade_data_processing.ipynb <==
import json
import os
import numpy as np

root = os.getcwd()
root = os.path.dirname(root)
import pandas as pd

os.chdir(os.path.join(root, "src", "utils"))
from aka import tri2name, tri2aka

covotes = pd.read_csv(os.path.join(root, "constructed_network", "covotes.csv"), index_col=0)
UN_countries = covotes.columns
# contravotes = pd.read_csv(os.path.join(root, "network_data", "contravotes.csv"), index_col=0)


# prefriendshipgraph = pd.read_csv(os.path.join(root, "network_data", "pre_friends_network_75.csv"), index_col=0)
# preenemyshipgraph = pd.read_csv(os.path.join(root, "network_data", "pre_enemies_network_75.csv"), index_col=0)
# alliancegraph = pd.read_csv(os.path.join(root, "network_data", "alliances_network.csv"), index_col=0)
# friendshipgraph = pd.read_csv(os.path.join(root, "network_data", "friends_network_75.csv"), index_col=0)
# enemyshipgraph  = pd.read_csv(os.path.join(root, "network_data", "enemies_network_75.csv"), index_col=0)


# pca_prefriends = pd.read_csv(os.path.join(root, "network_data", "pca_prefriends.csv"), index_col=0)
# pca_preenemies = pd.read_csv(os.path.join(root, "network_data", "pca_preenemies.csv"), index_col=0)
# pca_friends = pd.read_csv(os.path.join(root, "network_data", "pca_friends.csv"), index_col=0)
# pca_enemies  = pd.read_csv(os.path.join(root, "network_data", "pca_enemies.csv"), index_col=0)

# wto = pd.read_csv(os.path.join(root, "WTO_trade_data", "WTO_trade_data2.csv"), encoding='latin-1')
wto = pd.read_csv(os.path.join(root, "raw_data", "WTO_trade_data", "balanced_export.csv"), encoding='latin-1')

# set(wto["Partner Economy"])


from_countries = wto["Reporting Economy ISO3A Code"]
# countries = wto["Reporting Economy"]
to_countries = wto["Partner Economy ISO3A Code"]

from_to_countries = set(from_countries).union(to_countries)

from_to_countries = sorted(c for c in from_to_countries if not pd.isna(c))

from_to_countries

wto

wto.loc[100]

# try python loop. if too slow, then try other method
def process_row(row=wto.loc[100]):
    therefrom = row['Reporting Economy ISO3A Code']
    thereto = row['Partner Economy ISO3A Code']
    value = row['Value']
    if therefrom in bilateral.index and thereto in bilateral.columns:
        bilateral.loc[therefrom, thereto] = value


# pivot_df = wto.pivot_table(index='Reporting Economy', columns='Partner Economy', values='Value', aggfunc='sum', fill_value=0)
# pivot_df = wto.pivot_table(index='Reporting Economy ISO3A Code', columns='Partner Economy ISO3A Code', values='Value', aggfunc='sum', fill_value=0)

pivot_df2 = wto.pivot_table(index=["Reporting Economy ISO3A Code", "Partner Economy ISO3A Code", "Year"], values="Value", aggfunc='sum', fill_value=0)

pivot_df = pivot_df2.pivot_table(index="Reporting Economy ISO3A Code", columns=["Partner Economy ISO3A Code"], values="Value", aggfunc='mean', fill_value=0)

pivot_df

pivot_df.rename({"MMR":"MYR", "ROM":"ROU"}, axis=0, inplace=True)
pivot_df.rename({"MMR":"MYR", "ROM":"ROU"}, axis=1, inplace=True)

# missing countries
# for c in set(UN_countries).difference(set(pivot_df.index)):
#     print(tri2name[c])

# with open(os.path.join(root, "raw_data", "GDP_data", "GDP_data.json")) as f:
#     GDP = eval(f.read())

# # retrive the countries that doesn't have GDP data
# to_drop = []
# for c in pivot_df.columns:
#     try:
#         GDP[c]
#     except KeyError:
#         print(c)
#         to_drop.append(c)

# for c in to_drop:
#     pivot_df.drop(c, axis=1, inplace=True) if c in pivot_df.columns else None
#     pivot_df.drop(c, axis=0, inplace=True) if c in pivot_df.index   else None    

pivot_df.drop("VAT", axis=1, inplace=True) if "VAT" in pivot_df.columns else None
pivot_df.drop("EEC", axis=1, inplace=True) if "EEC" in pivot_df.columns else None
pivot_df.drop("CHT", axis=1, inplace=True) if "CHT" in pivot_df.columns else None

pivot_df.drop("VAT", axis=0, inplace=True) if "VAT" in pivot_df.index else None
pivot_df.drop("EEC", axis=0, inplace=True) if "EEC" in pivot_df.index else None
pivot_df.drop("CHT", axis=0, inplace=True) if "CHT" in pivot_df.index else None


# see how many empty values we got
print("total number of zero trade is ", (pivot_df == 0).sum().sum())
print("percentage-wise it is", (pivot_df == 0).sum().sum()/ (pivot_df.shape[0]**2)*100, "%")

# total = set(pivot_df.columns)
# given = set(pivot_df.index)
# unknown = total.difference(given)

# We only consider UN countries that are also in the WTO dataset
countries = sorted(set(UN_countries).intersection(set(pivot_df.columns)))
pivot_df = pivot_df.loc[countries, countries]

# check for zero trade flows. Only those involving PRK (north korea)
for c in countries:
    for cc in countries:
        if pivot_df.loc[c, cc] == 0:
            print(c, cc) if c < cc else None

def create_trade(pivot_df, perc, over=True):
    pivot_ratio_df = pivot_df.div(pivot_df.sum(axis=1), axis=0)
    mask = np.eye(pivot_ratio_df.shape[0], pivot_ratio_df.shape[1], dtype=bool)
    data = pivot_ratio_df.to_numpy()[~mask]
    threshold_percent = perc
    threshold = np.percentile(data, threshold_percent)
    if over:
        trade_overunder = pivot_ratio_df.applymap(lambda x: 1 if x>=threshold else 0)
    else:
        trade_overunder = pivot_ratio_df.applymap(lambda x:1 if x<threshold else 0)
    return trade_overunder

trade_over80 = create_trade(pivot_df, 80, over=True)
trade_over90 = create_trade(pivot_df, 90, over=True)

trade_under20 = create_trade(pivot_df, 20, over=False)
trade_under10 = create_trade(pivot_df, 10, over=False)


trans_trade_over80 = create_trade(pivot_df.T, 80, over=True)
trans_trade_over90 = create_trade(pivot_df.T, 90, over=True)

trans_trade_under20 = create_trade(pivot_df.T, 20, over=False)
trans_trade_under10 = create_trade(pivot_df.T, 10, over=False)

(pivot_df==0).sum().sum()

# wto

# make sure the trigraphs are consistent with ISO
# ISO3A = dict(zip(wto["Partner Economy ISO3A Code"], wto["Partner Economy"]))

# for tri, cname in ISO3A.items():
#     try:
#         print(f"ISO3A tri={tri}\nISO3A name  ={cname}\nmy trigraph ={tri2name[tri]}")
#     except KeyError:
#         print(f"{tri} is not found in my trigraphs----------------------------------")

## Result :
# CHT chinese taipei not present in UN country
# ROM romania in ISO code but ROU in my trigraph  => c
# EEC europeans union in ISO code but not in UN country

# pd.set_option('display.max_rows', 10)
# pivot_df.sum().sort_values(ascending=False)

# make a percentage-based pivot table.
# Each row is 
pivot_ratio_df = pivot_df.div(pivot_df.sum(axis=1), axis=0)

# import matplotlib.pyplot as plt
# import numpy as np

# percentiles_to_draw = [10, 20, 80, 90]

# # percentiles_to_draw = [10, 20, 30, 40]

# percentiles = np.percentile(pivot_ratio_df.to_numpy().flatten(), percentiles_to_draw)
# pivot_percentiles = dict(zip(percentiles_to_draw, percentiles))


# # Add vertical lines and annotations for percentiles
# for i, p in enumerate(percentiles):
#     plt.axvline(p, color='r', linestyle='--', linewidth=2)
#     plt.text(p, 50, f'{percentiles_to_draw[i]}\% {p:.2f}', color='r', rotation=90, verticalalignment='bottom')

# # Add labels and title
# plt.xlabel('Value')
# plt.ylabel('Frequency')

# show_up_to_percentile= 95

# # plot histogram up to certain percentile
# data = pivot_ratio_df.to_numpy()
# mask = np.eye(data.shape[0], data.shape[1], dtype=bool)
# data = data[~mask]
# filtered_data = data[data <= np.percentile(data, show_up_to_percentile)]
# plt.hist(filtered_data, bins=15)

# plt.title('')


# def make_symmetric(df):
#     return (df + df.T)>=1

# threshold_percent = 90
# threshold = np.percentile(data, threshold_percent)
# trade_over_90 = pivot_ratio_df.applymap(lambda x: 1 if x>=threshold else 0)


# threshold_percent = 80
# threshold = np.percentile(data, threshold_percent)
# trade_over_80 = pivot_ratio_df.applymap(lambda x: 1 if x>=threshold else 0)
# print(data.mean())


# threshold_percent = 75
# threshold = np.percentile(data, threshold_percent)
# trade_over_75 = pivot_ratio_df.applymap(lambda x: 1 if x>=threshold else 0)




# threshold_percent = 25
# threshold = np.percentile(data, threshold_percent)
# trade_under_25 = pivot_ratio_df.applymap(lambda x: 1 if x<threshold else 0)

# threshold_percent = 20
# threshold = np.percentile(data, threshold_percent)
# trade_under_20 = pivot_ratio_df.applymap(lambda x: 1 if x<threshold else 0)


# threshold_percent = 10
# threshold = np.percentile(data, threshold_percent)
# trade_under_10 = pivot_ratio_df.applymap(lambda x: 1 if x<threshold else 0)


# how many partnar countries does each reporting country has?
# plt.hist(trade_over_80.sum(axis=1))
# CCL : about 15-20 most likely. looks like noramlly distributed.

trade_over90.to_csv(os.path.join(root, "constructed_network", "export_over90.csv"))
trade_over80.to_csv(os.path.join(root, "constructed_network", "export_over80.csv"))
# trade_over_75.to_csv(os.path.join(root, "constructed_network", "export_over75.csv"))

# trade_under_25.to_csv(os.path.join(root,"constructed_network",  "export_under25.csv"))
trade_under20.to_csv(os.path.join(root,"constructed_network",  "export_under20.csv"))
trade_under10.to_csv(os.path.join(root,"constructed_network",  "export_under10.csv"))


trans_trade_over90.to_csv(os.path.join(root, "constructed_network", "import_over90.csv"))
trans_trade_over80.to_csv(os.path.join(root, "constructed_network", "import_over80.csv"))
# trade_over_75.to_csv(os.path.join(root, "constructed_network", "export_over75.csv"))

# trade_under_25.to_csv(os.path.join(root,"constructed_network",  "export_under25.csv"))
trans_trade_under20.to_csv(os.path.join(root,"constructed_network",  "import_under20.csv"))
trans_trade_under10.to_csv(os.path.join(root,"constructed_network",  "import_under10.csv"))
