==> src/10. Network_visualization.ipynb <==
# !git clone -b master https://github.com/StefanoNasini/Country-by-country-2022-2023-international-relationships
# !cd Country-by-country-2022-2023-international-relationships

import json
import os

root = os.path.dirname(os.getcwd())
import pandas as pd

import numpy as np
import networkx as nx
import matplotlib.pyplot as plt

def network_plot(adj_matrix, layout=nx.kamada_kawai_layout, width=0.3):
    # Create a graph from the adjacency matrix
    try:
        adj_matrix = adj_matrix.to_numpy()
    except _:
        pass
    G = nx.from_numpy_array(adj_matrix)

    # Plot the largest component
    plt.figure(figsize=(6, 6))
    pos = layout(G)  # You can change the layout algorithm as needed
    nx.draw(G, pos, with_labels=False, node_size=20, node_color='black', font_size=10, font_weight='bold', edge_color='gray', width=width)
    # plt.title("Largest Component of the Graph")

    num_nodes = G.number_of_nodes()
    num_edges = G.number_of_edges()
    # caption = f"Number of Nodes: {num_nodes}\nNumber of Edges: {num_edges}"
    # plt.text(0.5, -0.1, caption, transform=plt.gca().transAxes, fontsize=12, ha='center')
    print(num_nodes, num_edges)
    plt.show()

def largest_component_plot(adj_matrix, layout=nx.kamada_kawai_layout, width=0.3, save="", format_=""):
    
    try:
        adj_matrix = adj_matrix.to_numpy()
    except _:
        pass
    
    # Create a graph from the adjacency matrix
    for i in range(adj_matrix.shape[0]):
        adj_matrix[i, i] = 0
        
    G = nx.from_numpy_array(adj_matrix)


    # Find connected components and get the largest component
    largest_component = max(nx.connected_components(G), key=len)
    largest_component_subgraph = G.subgraph(largest_component)

    # Plot the largest component
    plt.figure(figsize=(6, 6))
    pos = nx.kamada_kawai_layout(largest_component_subgraph)  # You can change the layout algorithm as needed
    nx.draw(largest_component_subgraph, pos, with_labels=False, node_size=20, node_color='black', font_size=10, font_weight='bold', edge_color='gray', width=width)
    # plt.title("Largest Component of the Graph")

    num_nodes = largest_component_subgraph.number_of_nodes()
    num_edges = largest_component_subgraph.number_of_edges()
    # caption = f"Number of Nodes: {num_nodes}\nNumber of Edges: {num_edges}"
    # plt.text(0.5, -0.1, caption, transform=plt.gca().transAxes, fontsize=12, ha='center')
    print(num_nodes, num_edges)

    if save=="":
        plt.show()
    elif format_ == 'pdf':
        plt.savefig(save, format = "pdf")
        plt.close()
    else:
        plt.savefig(save)
        plt.close()
    

import networkx as nx
import matplotlib.pyplot as plt

def draw_two_networks(adj_matrix_1, adj_matrix_2, layout=nx.kamada_kawai_layout, width=0.3, edge_color_1='gray', edge_color_2='blue', save="", format_=""):
    # Create graphs from the adjacency matrices
    try:
        adj_matrix_1 = adj_matrix_1.to_numpy()
    except AttributeError:
        pass
    try:
        adj_matrix_2 = adj_matrix_2.to_numpy()
    except AttributeError:
        pass
    
    for i in range(adj_matrix_1.shape[0]):
        adj_matrix_1[i, i] = 0
        
    for i in range(adj_matrix_2.shape[0]):
        adj_matrix_2[i, i] = 0
    
    G1 = nx.from_numpy_array(adj_matrix_1)
    G2 = nx.from_numpy_array(adj_matrix_2)
    
    friend_edges = list(G1.edges())
    enemy_edges = list(G2.edges())

    G = nx.Graph()

    # Add friend edges with 'relationship' attribute set to 'friend'
    for edge in friend_edges:
        G.add_edge(edge[0], edge[1], relationship='friend')

    # Add enemy edges with 'relationship' attribute set to 'enemy'
    for edge in enemy_edges:
        G.add_edge(edge[0], edge[1], relationship='enemy')

    # Plot the graph with different edge colors based on the 'relationship' attribute
    plt.figure(figsize=(8, 6))

    pos = layout(G)
    friend_edges = [(u, v) for (u, v, rel) in G.edges(data='relationship') if rel == 'friend']
    enemy_edges = [(u, v) for (u, v, rel) in G.edges(data='relationship') if rel == 'enemy']

    nx.draw(G, pos, with_labels=False, node_size=20, node_color='black', font_size=12, font_weight='bold', width=width, edge_color='gray')
    nx.draw_networkx_edges(G, pos, edgelist=friend_edges, edge_color='green', width=width)
    nx.draw_networkx_edges(G, pos, edgelist=enemy_edges, edge_color='red', width=width)

    num_nodes = G.number_of_nodes()
    num_edges = G.number_of_edges()
    print(num_nodes, num_edges)
    
    if save == "":
        plt.show()
    elif format_ == 'pdf':
        plt.savefig(save, format = "pdf")
        plt.close()
    else:
        plt.savefig(save)
        plt.close()
    

# largest_component_plot(prefriendshipgraph, layout=nx.kamada_kawai_layout, width=0.3)
# # network_plot(prefriendshipgraph, layout=nx.random_layout, width=0.3)

# largest_component_plot(preenemyshipgraph, layout=nx.kamada_kawai_layout, width=0.6)

# # new one 0208
# largest_component_plot(alliancegraph, layout=nx.kamada_kawai_layout, width=0.6 )
# network_plot(alliancegraph, layout=nx.kamada_kawai_layout, width=0.6 )


# largest_component_plot(alliancegraph, layout=nx.kamada_kawai_layout, width=0.6 )
# network_plot(alliancegraph, layout=nx.kamada_kawai_layout, width=0.6 )

# largest_component_plot(friendshipgraph, layout=nx.kamada_kawai_layout, width=0.3)
# network_plot(friendshipgraph, layout=nx.random_layout, width=0.3)

# # 0208
# largest_component_plot(friendshipgraph, layout=nx.kamada_kawai_layout, width=0.3)
# network_plot(friendshipgraph, layout=nx.random_layout, width=0.3)

# largest_component_plot(enemyshipgraph, layout=nx.kamada_kawai_layout, width=0.3)
# network_plot(enemyshipgraph, layout=nx.kamada_kawai_layout, width=0.3)

# # 0208
# largest_component_plot(enemyshipgraph, layout=nx.kamada_kawai_layout, width=0.3)
# network_plot(enemyshipgraph, layout=nx.kamada_kawai_layout, width=0.3)

# draw_two_networks(friendshipgraph, enemyshipgraph, layout=nx.kamada_kawai_layout, width=0.2, edge_color_1='green', edge_color_2='red')


# #0208

# draw_two_networks(friendshipgraph, enemyshipgraph, layout=nx.kamada_kawai_layout, width=0.2, edge_color_1='green', edge_color_2='red')


# pca_prefriends = pd.read_csv(os.path.join(root, "network_data", "pca_prefriends.csv"), index_col=0)
# pca_preenemies = pd.read_csv(os.path.join(root, "network_data", "pca_preenemies.csv"), index_col=0)
# pca_friends = pd.read_csv(os.path.join(root, "network_data", "pca_friends.csv"), index_col=0)
# pca_enemies  = pd.read_csv(os.path.join(root, "network_data", "pca_enemies.csv"), index_col=0)

# largest_component_plot(pca_prefriends, layout=nx.kamada_kawai_layout, width=0.3)


# largest_component_plot(pca_friends, layout=nx.kamada_kawai_layout, width=0.3)


# largest_component_plot(pca_preenemies, layout=nx.kamada_kawai_layout, width=0.3)


# largest_component_plot(pca_enemies, layout=nx.kamada_kawai_layout, width=0.3)


# draw_two_networks(pca_friends, pca_enemies , layout=nx.spring_layout, width=0.2, edge_color_1='green', edge_color_2='red')



covotes = pd.read_csv(os.path.join(root, "constructed_network", "covotes.csv"), index_col=0)
contravotes = pd.read_csv(os.path.join(root, "constructed_network", "contravotes.csv"), index_col=0)

covotes_over80 = pd.read_csv(os.path.join(root, "constructed_network", "cocontravotes", "covotes_over80.csv"), index_col=0)
covotes_over90 = pd.read_csv(os.path.join(root, "constructed_network", "cocontravotes", "covotes_over90.csv"), index_col=0)
covotes_under20 = pd.read_csv(os.path.join(root, "constructed_network", "cocontravotes", "covotes_under20.csv"), index_col=0)
covotes_under10 = pd.read_csv(os.path.join(root, "constructed_network", "cocontravotes", "covotes_under10.csv"), index_col=0)
contravotes_under20 = pd.read_csv(os.path.join(root, "constructed_network", "cocontravotes", "contravotes_under20.csv"), index_col=0)
contravotes_under10 = pd.read_csv(os.path.join(root, "constructed_network", "cocontravotes", "contravotes_under10.csv"), index_col=0)
contravotes_over80 = pd.read_csv(os.path.join(root, "constructed_network", "cocontravotes", "contravotes_over80.csv"), index_col=0)
contravotes_over90 = pd.read_csv(os.path.join(root, "constructed_network", "cocontravotes", "contravotes_over90.csv"), index_col=0)


alliances = pd.read_csv(os.path.join(root, "constructed_network", "alliances.csv"), index_col=0)
alliances.drop(columns="MAC", inplace=True)

alliances_zero = 1-alliances

trade_over80 = pd.read_csv(os.path.join(root, "constructed_network", "export_over80.csv"), index_col=0)
trade_over90 = pd.read_csv(os.path.join(root, "constructed_network", "export_over90.csv"), index_col=0)

trade_under20 = pd.read_csv(os.path.join(root, "constructed_network", "export_under20.csv"), index_col=0)
trade_under10 = pd.read_csv(os.path.join(root, "constructed_network", "export_under10.csv"), index_col=0)

trans_trade_over80 = pd.read_csv(os.path.join(root, "constructed_network", "import_over80.csv"), index_col=0)
trans_trade_over90 = pd.read_csv(os.path.join(root, "constructed_network", "import_over90.csv"), index_col=0)

trans_trade_under20 = pd.read_csv(os.path.join(root, "constructed_network", "import_under20.csv"), index_col=0)
trans_trade_under10 = pd.read_csv(os.path.join(root, "constructed_network", "import_under10.csv"), index_col=0)



friends80_bi = pd.read_csv(os.path.join(root, "constructed_network", "final_network", "friends80_bi.csv"), index_col=0)
friends80_tri = pd.read_csv(os.path.join(root,"constructed_network",  "final_network", "friends80_tri.csv"), index_col=0)
friends90_bi = pd.read_csv(os.path.join(root, "constructed_network", "final_network", "friends90_bi.csv"), index_col=0)
friends90_tri = pd.read_csv(os.path.join(root,"constructed_network",  "final_network", "friends90_tri.csv"), index_col=0)
enemies80_bi = pd.read_csv(os.path.join(root, "constructed_network", "final_network", "enemies80_bi.csv"), index_col=0)
enemies80_tri = pd.read_csv(os.path.join(root,"constructed_network",  "final_network", "enemies80_tri.csv"), index_col=0)
enemies90_bi = pd.read_csv(os.path.join(root, "constructed_network", "final_network", "enemies90_bi.csv"), index_col=0)
enemies90_tri = pd.read_csv(os.path.join(root,"constructed_network",  "final_network", "enemies90_tri.csv"), index_col=0)


networks = ['alliances', 'covotes_over80', 'covotes_over90', 'covotes_under20', 'covotes_under10', 
            'contravotes_under20', 'contravotes_under10', 'contravotes_over80', 'contravotes_over90', 
            'trade_over80', 'trade_over90', 'trade_under20', 'trade_under10',
            'trans_trade_over80', 'trans_trade_over90', 'trans_trade_under20', 'trans_trade_under10',
            'friends80_bi', 'friends90_bi', 'enemies80_bi', 'enemies90_bi', 
            'friends80_tri', 'friends90_tri', 'enemies80_tri', 'enemies90_tri']


# networks = ['trade_over80', 'trade_over90', 'trade_under20', 'trade_under10']

# save as pdf files
for network in networks:
    df = eval(network)
    if not os.path.exists(os.path.join(root, "constructed_network", "graph")):
        os.mkdir(os.path.join(root, "constructed_network", "graph"))
    savename = os.path.join(root, "constructed_network", "graph", "pdfs", f"{network}.pdf")
    print(f'Plotting {network}. Its (n_nodes, n_edges) are ')
    largest_component_plot(df, layout=nx.kamada_kawai_layout, width=0.2, save=savename, format_="pdf")
    print()

# save as png files
for network in networks:
    df = eval(network)
    if not os.path.exists(os.path.join(root, "constructed_network", "graph", "pngs")):
        os.mkdir(os.path.join(root, "constructed_network", "graph", "pngs"))
    savename = os.path.join(root, "constructed_network", "graph", "pngs", f"{network}.png")
    print(f'Plotting {network}. Its (n_nodes, n_edges) are ')
    largest_component_plot(df, layout=nx.kamada_kawai_layout, width=0.2, save=savename)
    print()

networks

endings = "80_bi 90_bi 80_tri 90_tri"
# save as png files
for network in endings.split(' '):
    friend = "friends"+ network
    enemy = "enemies" + network
    df_friend, df_enemy = eval(friend), eval(enemy)
    if not os.path.exists(os.path.join(root, "constructed_network", "graph", "pngs")):
        os.mkdir(os.path.join(root, "constructed_network", "graph", "pngs"))
    savename = os.path.join(root, "constructed_network", "graph", "pngs", f"combined{network}.png")
    print(f'Plotting {network}. Its (n_nodes, n_edges) are ')
    draw_two_networks(df_friend, df_enemy, layout=nx.kamada_kawai_layout, width=0.2, edge_color_1='green', edge_color_2='red', save=savename, format_="png") 

endings = "80_bi 90_bi 80_tri 90_tri"
# save as pdf files
for network in endings.split(' '):
    friend = "friends"+ network
    enemy = "enemies" + network
    df_friend, df_enemy = eval(friend), eval(enemy)
    if not os.path.exists(os.path.join(root, "constructed_network", "graph", "pdfs")):
        os.mkdir(os.path.join(root, "constructed_network", "graph", "pdfs"))
    savename = os.path.join(root, "constructed_network", "graph", "pdfs", f"combined{network}.pdf")
    print(f'Plotting {network}. Its (n_nodes, n_edges) are ')
    draw_two_networks(df_friend, df_enemy, layout=nx.kamada_kawai_layout, width=0.2, edge_color_1='green', edge_color_2='red', save=savename, format_="pdf")
