==> src/11. distances_between_countries.ipynb <==
# !git clone -b master https://github.com/StefanoNasini/Country-by-country-2022-2023-international-relationships
# !cd Country-by-country-2022-2023-international-relationships

import json
import os

root = os.path.dirname(os.getcwd())
import pandas as pd

os.chdir(os.path.join(root, "src", "utils"))
from aka import tri2name, tri2aka
os.chdir(os.path.join(root, "src"))

covotes = pd.read_csv(os.path.join(root, "constructed_network", "covotes.csv"), index_col=0)
UN_countries = covotes.columns

with open(os.path.join(root, "raw_data", "country_distance_data", "country_dist.txt")) as f:
    l = f.readlines()
    
with open(os.path.join(root, "raw_data", "country_distance_data", "country_border.txt")) as f:
    border = f.readlines()
    


ll = []
tri = set()
for e in l[1:]:
    ll.append(e.split(' '))
    tri.add(ll[-1][2][1:-1])
    tri1 = ll[-1][1][1:-1]
    tri2 = ll[-1][2][1:-1]
    dist = ll[-1][3][1:-2]

import pandas as pd
dists = pd.DataFrame(columns=sorted(UN_countries), index=sorted(UN_countries))
dists.iloc[:, :] = float("nan")

df_border = pd.DataFrame(columns=sorted(UN_countries), index=sorted(UN_countries), dtype=bool)
df_border.iloc[:, :] = False




for e in l[1:]:
    row = (e.split(' '))
    tri.add(row[2][1:-1])
    tri1 = row[1][1:-1]
    tri2 = row[2][1:-1]
    dist = row[3][1:-2]
#     if tri1 in dists.index and tri2 in dists.columns:
    if True:
        dists.loc[tri1, tri2] = float(dist)

to_be_filled = dists[dists.isna().sum()!=9].index
to_be_filled

# Use neighbor country to replace the distance to the missing data
tritos= "COG, CHE, FRA, YUG, LAO, ROM, YUG, SDN, IDN".split(', ')

for (trifrom, trito) in zip(to_be_filled, tritos):
    dists.loc[:, trifrom] = dists.loc[:, trito]
    dists.loc[trifrom, :] = dists.loc[trito, :]

dists

for tri_ in tri:
    try:
        
        (tri2aka[tri_])
    except KeyError:
        print(tri_)
        
# west saraha
# zar : democratic congo
# YUG : serbia & montenegro
# liechtenstein : use switzerland
# myanmar : use laos
# south sudan : use sudan

border

ll = []
for e in border[1:]:
#     ll.append(e.split(' '))
#     tri.add(ll[-1][2][1:-1])
    row = e.split(' ')
    tri1 = row[1][1:-1]
    tri2 = row[2][1:-1]
    is_border = row[3][1:-2]
    if is_border == '1':
        df_border.loc[tri1, tri2] = True
    
# #     if tri1 in dists.index and tri2 in dists.columns:
#     dists.loc[tri1, tri2] = float(dist)

df_border

for t in set(UN_countries).difference(tri):
    print(tri2name[t], t)

tri2aka

# construct the dij matrix. dij = 1 if neighbor else dmax-dij / dmax
dij = pd.DataFrame(columns=sorted(UN_countries), index=sorted(UN_countries))
dij.iloc[:, :] = float("nan")

dmax = dists.max().max()

for tri1 in UN_countries:
    for tri2 in UN_countries:
        if df_border.loc[tri1, tri2] == True:
            dij.loc[tri1, tri2] = 1
        else:
            dij.loc[tri1, tri2] = (dmax-dists.loc[tri1, tri2]) / dmax

dij.to_csv(os.path.join(root, "constructed_network", "distance_factor.csv"))
